+++ lib/flop_phoenix/pagination.ex
@@ -24,36 +24,11 @@ defmodule Flop.Phoenix.Pagination do
   end
 
   def build_page_link_helper(meta, route_helper, route_helper_args) do
-    filter_map =
-      meta.flop.filters
-      |> Stream.with_index()
-      |> Enum.into(%{}, fn {filter, index} ->
-        {index, Map.from_struct(filter)}
-      end)
-
-    query_params =
-      []
-      |> maybe_add_param(:filters, filter_map)
-      |> maybe_add_param(:order_by, meta.flop.order_by)
-      |> maybe_add_param(:order_directions, meta.flop.order_directions)
-      |> maybe_add_param(:page_size, meta.page_size)
-
-    [last_arg | rest] = Enum.reverse(route_helper_args)
+    query_params = Flop.Phoenix.to_query(meta.flop)
 
     fn page ->
-      final_route_helper_args =
-        if is_list(last_arg) do
-          query_arg =
-            last_arg
-            |> Keyword.merge(query_params)
-            |> Keyword.put(:page, page)
-
-          Enum.reverse([query_arg | rest])
-        else
-          route_helper_args ++ [Keyword.put(query_params, :page, page)]
-        end
-
-      apply(route_helper, final_route_helper_args)
+      params = Keyword.put(query_params, :page, page)
+      Flop.Phoenix.build_path(route_helper, route_helper_args, params)
     end
   end
 
@@ -69,14 +44,8 @@ defmodule Flop.Phoenix.Pagination do
     if meta.has_previous_page? do
       attrs = Keyword.put(attrs, :to, page_link_helper.(meta.previous_page))
 
-      if opts[:live_view] do
-        live_patch attrs do
-          content
-        end
-      else
-        link attrs do
-          content
-        end
+      live_patch attrs do
+        content
       end
     else
       attrs = Keyword.put(attrs, :disabled, "disabled")
@@ -99,14 +68,8 @@ defmodule Flop.Phoenix.Pagination do
     if meta.has_next_page? do
       attrs = Keyword.put(attrs, :to, page_link_helper.(meta.next_page))
 
-      if opts[:live_view] do
-        live_patch attrs do
-          content
-        end
-      else
-        link attrs do
-          content
-        end
+      live_patch attrs do
+        content
       end
     else
       attrs = Keyword.put(attrs, :disabled, "disabled")
@@ -168,15 +131,9 @@ defmodule Flop.Phoenix.Pagination do
         do: pagination_ellipsis(ellipsis_class, ellipsis_content),
         else: raw(nil)
 
-    link_func =
-      if opts[:live_view],
-        do: &live_patch/2,
-        else: &link/2
-
     first_link =
       if first > 1,
-        do:
-          page_link_tag(1, meta, link_attrs, aria_label, route_func, link_func),
+        do: page_link_tag(1, meta, link_attrs, aria_label, route_func),
         else: raw(nil)
 
     last_link =
@@ -187,14 +144,13 @@ defmodule Flop.Phoenix.Pagination do
             meta,
             link_attrs,
             aria_label,
-            route_func,
-            link_func
+            route_func
           ),
         else: raw(nil)
 
     links =
       for page <- range do
-        page_link_tag(page, meta, link_attrs, aria_label, route_func, link_func)
+        page_link_tag(page, meta, link_attrs, aria_label, route_func)
       end
 
     content_tag :ul, list_attrs do
@@ -213,8 +169,7 @@ defmodule Flop.Phoenix.Pagination do
          meta,
          link_attrs,
          aria_label,
-         route_func,
-         link_func
+         route_func
        ) do
     attrs =
       link_attrs
@@ -226,7 +181,7 @@ defmodule Flop.Phoenix.Pagination do
       |> Keyword.put(:to, route_func.(page))
 
     content_tag :li do
-      link_func.(page, attrs)
+      live_patch(page, attrs)
     end
   end
 
@@ -263,8 +218,4 @@ defmodule Flop.Phoenix.Pagination do
     |> Keyword.update!(:aria, &Keyword.put(&1, :current, "page"))
     |> Keyword.update!(:class, &"#{&1} is-current")
   end
-
-  defp maybe_add_param(params, _, nil), do: params
-  defp maybe_add_param(params, _, []), do: params
-  defp maybe_add_param(params, key, value), do: Keyword.put(params, key, value)
 end